Tag all bugs with the team
This tweaks the bugstore a bit so that it knows about "a lot of bugs"
but the queries all return sets of bugs for a specified team. This fixes
the issue where adding blocker bugs to the bugMap added non-team bugs to
the bugMap.

It also fixes the problem where switching between teams caused you to see
a union of the bug sets.

Fixes #61